L-11 modified app,config,wsgi added comands
