+++ blog/comands.py
@@ -1,0 +1,38 @@
+import os
+from werkzeug.security import generate_password_hash
+from wsgi import app
+from blog.models.database import db
+
+@app.cli.command("create-admin")
+def create_users():
+    """
+    Run in your terminal:
+    flask create-admin
+    > done! created users: <User #1 'admin'> <User #2 'james'>
+    """
+    from blog.models import User
+    admin = User(username="admin", is_staff=True, email='dfsfbcv@email.com')
+    admin.password = os.environ.get("ADMIN_PASSWORD") or "adminpass"
+
+    db.session.add(admin)
+    db.session.commit()
+    print("done! created admin:", admin)
+
+@app.cli.command("create-tags")
+def create_tags():
+    """
+    Run in your terminal:
+    ➜ flask create-tags
+    """
+    from blog.models import Tag
+    for name in [
+        "flask",
+        "django",
+        "python",
+        "sqlalchemy",
+        "news",
+    ]:
+        tag = Tag(name=name)
+        db.session.add(tag)
+    db.session.commit()
+    print("created tags")
